Firebase Stub: unit tests separated in different files
